I am done for now on making auto makefile
SRCS is wokring as expected now, the only thing
I would like to change is to get rid of the
dependency of the python file to include the
recipie needed. If I can package it into the Makefile
this would be ideal! Overall though, I am happy with th
resulting Makefile