OK, now I think I'm done with this Makefile
I think that this makefile is now good enough to make
any of my small scale projects! I added a release and
debug target, and everything is all smooth sailing
Python script is a bit ugly, and is something I will
be thinking to improve. But the makefile is all in one!