Obviously can make better, but good enough
I am thinking about extending the struct Term with a core
struct, in order to allow for terms that are composed of
more than one variable. Basically you would  have a vector
or unordered map of core_terms, though it is to note, that
this would make the implentation of the whole program
significantly harder to still make efficient.